backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
httpx
python-multipart
PyNaCl
